pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voiptest"
version = "0.1.0"
description = "VoIP regression smoke tests for CI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "VoIP Test Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
voiptest = "voiptest.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["voiptest*"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"voiptest.engines" = [
    "sipp_scenarios/*.xml",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
